Move MemoryJmapSettingsRepositoryModule to apps/memory module
